fix(ci): extract convertSpeed to own file to avoid coverage regression
Moving convertSpeed to speedConversion.ts prevents mapHelpers.tsx
(12% coverage, mostly React rendering) from being loaded during tests,
which was dragging src/utils/** below the 70% threshold.
