+++ src/utils/mapHelpers.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 import L from 'leaflet';
 import { Marker, Tooltip, Popup } from 'react-leaflet';
 import { PositionHistoryItem } from '../contexts/MapContext';
+import { convertSpeed } from './speedConversion';
+export { convertSpeed };
 
 /**
  * Scaled SNR sentinel for MQTT/unknown hops.
@@ -20,17 +22,6 @@ export const MQTT_SNR_ZERO = 0;
 /** Returns true if the scaled SNR value indicates an MQTT/unknown hop */
 export const isMqttSnr = (snr: number | undefined): boolean =>
   snr === MQTT_SNR_SENTINEL || snr === MQTT_SNR_ZERO;
-
-/**
- * Convert ground speed from m/s (Meshtastic protobuf uint32) to display units.
- * Returns { speed, unit } where speed is rounded to 1 decimal place.
- */
-export function convertSpeed(metersPerSecond: number, distanceUnit: string): { speed: number; unit: string } {
-  const speedKmh = metersPerSecond * 3.6;
-  const speed = distanceUnit === 'mi' ? speedKmh * 0.621371 : speedKmh;
-  const unit = distanceUnit === 'mi' ? 'mph' : 'km/h';
-  return { speed: parseFloat(speed.toFixed(1)), unit };
-}
 
 // Constants for arrow generation
 const ARROW_DISTANCE_THRESHOLD = 0.05; // One arrow per 0.05 degrees
